Move to #core/dom/shadow-embed
