+++ src/preact/base-element.js
@@ -19,8 +19,8 @@ import {
 } from '#core/dom/layout';
 import {MediaQueryProps} from '#core/dom/media-query-props';
 import {childElementByAttr, childElementByTag} from '#core/dom/query';
+import {installShadowStyle} from '#core/dom/shadow-embed';
 import {PauseHelper} from '#core/dom/video/pause-helper';
-import {installShadowStyle} from '#core/dom/web-components';
 import * as mode from '#core/mode';
 import {isElement} from '#core/types';
 import {dict, hasOwn, map} from '#core/types/object';
